migrate completions to another file
