+++ src/completions.ts
@@ -1,0 +1,96 @@
+import * as vscode from 'vscode';
+import { declarationsAndTargets , dataformTags } from './extension';
+
+
+export const sourcesAutoCompletionDisposable = () => vscode.languages.registerCompletionItemProvider(
+    // NOTE: Could this be made more reusable, i.e. a function that takes in the trigger and the language
+    /*
+    you might need to set up the file association to use the auto-completion
+    sql should be added as a file association for sqlx
+    this will enable both sufficient syntax highlighting and auto-completion
+    */
+    { language: 'sql', scheme: 'file' },
+    {
+        provideCompletionItems(document, position, token, context) {
+
+            const linePrefix = document.lineAt(position).text.substring(0, position.character);
+            if (!linePrefix.endsWith('$')) {
+                return undefined;
+            }
+            let sourceCompletionItem = (text: any) => {
+                let item = new vscode.CompletionItem(text, vscode.CompletionItemKind.Field);
+                item.range = new vscode.Range(position, position);
+                return item;
+            };
+            if (declarationsAndTargets.length === 0) {
+                return undefined;
+            }
+            let sourceCompletionItems: vscode.CompletionItem[] = [];
+            declarationsAndTargets.forEach((source: string) => {
+                source = `{ref("${source}")}`;
+                sourceCompletionItems.push(sourceCompletionItem(source));
+            });
+            return sourceCompletionItems;
+        }
+    },
+    '$' // trigger
+);
+
+export const dependenciesAutoCompletionDisposable = () => vscode.languages.registerCompletionItemProvider(
+    // NOTE: Could this be made more reusable, i.e. a function that takes in the trigger and the language
+    { language: 'sql', scheme: 'file' },
+    {
+        provideCompletionItems(document, position, token, context) {
+
+            const linePrefix = document.lineAt(position).text.substring(0, position.character);
+            if (!linePrefix.includes('dependencies')) {
+                return undefined;
+            }
+            let sourceCompletionItem = (text: any) => {
+                let item = new vscode.CompletionItem(text, vscode.CompletionItemKind.Field);
+                item.range = new vscode.Range(position, position);
+                return item;
+            };
+            if (declarationsAndTargets.length === 0) {
+                return undefined;
+            }
+            let sourceCompletionItems: vscode.CompletionItem[] = [];
+            declarationsAndTargets.forEach((source: string) => {
+                source = `${source}`;
+                sourceCompletionItems.push(sourceCompletionItem(source));
+            });
+            return sourceCompletionItems;
+        },
+    },
+    ...["'", '"'],
+);
+
+export const tagsAutoCompletionDisposable = () => vscode.languages.registerCompletionItemProvider(
+    // NOTE: Could this be made more reusable, i.e. a function that takes in the trigger and the language
+    { language: 'sql', scheme: 'file' },
+    {
+        provideCompletionItems(document, position, token, context) {
+
+            const linePrefix = document.lineAt(position).text.substring(0, position.character);
+            if (!linePrefix.includes('tags')) {
+                return undefined;
+            }
+            let sourceCompletionItem = (text: any) => {
+                let item = new vscode.CompletionItem(text, vscode.CompletionItemKind.Field);
+                item.range = new vscode.Range(position, position);
+                return item;
+            };
+            if (dataformTags.length === 0) {
+                return undefined;
+            }
+            let sourceCompletionItems: vscode.CompletionItem[] = [];
+            dataformTags.forEach((source: string) => {
+                source = `${source}`;
+                sourceCompletionItems.push(sourceCompletionItem(source));
+            });
+            return sourceCompletionItems;
+        },
+    },
+    ...["'", '"'],
+);
+
